Split the screenshot harness and lift the sync contracts out of the coordinator
Screenshot.kt held the offscreen render entry point together with every seeded
fixture and fake transport it needs:

  Screenshot            950 -> 244 + seeds 414 + fakes 151 + previews 171

SyncCoordinator's file opened with ten pure contract and domain types before the
class itself. Those move to SyncContracts.kt (1906 -> 1739). The coordinator
class is untouched: its 1692 lines still need decomposing, and that means
deciding who owns the session and the sync mutex first.

The detekt baseline is regenerated after the moves: 122 entries down to 116,
with nothing in it that was not already suppressed before this branch.
