+++ composeApp/src/desktopMain/kotlin/app/skerry/ui/desktop/Screenshot.kt
@@ -1,99 +1,23 @@
 package app.skerry.ui.desktop
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.ImageComposeScene
 import androidx.compose.ui.unit.Density
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import app.skerry.shared.files.FileContentBrowser
-import app.skerry.shared.files.FileItem
-import app.skerry.shared.files.FileItemType
-import app.skerry.shared.host.Host
-import app.skerry.shared.ssh.ConnectionType
 import app.skerry.ui.host.HostSection
-import app.skerry.ui.files.FileEditController
-import app.skerry.ui.files.FileEditorScreen
-import app.skerry.shared.host.HostStore
-import app.skerry.shared.terminal.Asciicast
-import app.skerry.shared.terminal.CastEvent
 import app.skerry.ui.terminal.TerminalThemes
-import app.skerry.shared.sftp.SftpClient
-import app.skerry.shared.sftp.SftpEntry
-import app.skerry.shared.sftp.SftpEntryType
-import app.skerry.shared.sftp.SftpProgress
-import app.skerry.shared.ssh.DynamicForwardSpec
-import app.skerry.shared.ssh.ExecResult
-import app.skerry.shared.ssh.LocalForwardSpec
-import app.skerry.shared.ssh.PortForward
-import app.skerry.shared.ssh.PtySize
-import app.skerry.shared.ssh.RemoteForwardSpec
-import app.skerry.shared.ssh.ShellChannel
-import app.skerry.shared.ssh.SshAuth
-import app.skerry.shared.ssh.HostKeyMismatch
-import app.skerry.shared.ssh.HostKeyMismatchStore
-import app.skerry.shared.ssh.KnownHost
-import app.skerry.shared.ssh.KnownHostsStore
-import app.skerry.shared.ssh.SshConnection
-import app.skerry.shared.ssh.SshTarget
-import app.skerry.shared.ssh.SshTransport
-import app.skerry.shared.tunnel.Tunnel
-import app.skerry.shared.tunnel.TunnelDirection
-import app.skerry.shared.tunnel.TunnelStore
-import app.skerry.ui.tunnel.SERVICE_SCAN_COMMAND
-import app.skerry.ui.tunnel.TunnelManager
-import app.skerry.ui.tunnel.TunnelTelemetry
-import app.skerry.ui.tunnel.TunnelStatus
-import app.skerry.ui.tunnel.TunnelResolution
-import app.skerry.ui.tunnel.TunnelUnavailable
 import app.skerry.shared.vault.BouncyCastleSshKeyGenerator
 import app.skerry.shared.vault.SshjCertificateInspector
-import app.skerry.shared.vault.CredentialStore
-import app.skerry.shared.vault.DataKey
-import app.skerry.shared.vault.MergeResult
-import app.skerry.shared.vault.RecordType
-import app.skerry.shared.vault.SshKeyGenerator
-import app.skerry.shared.vault.SyncMeta
-import app.skerry.shared.vault.SshKeyType
-import app.skerry.shared.vault.UnlockResult
 import app.skerry.shared.vault.Vault
-import app.skerry.shared.vault.VaultRecord
 import app.skerry.ui.AppDependencies
-import app.skerry.ui.identity.CredentialDraft
-import app.skerry.ui.identity.CredentialKind
 import app.skerry.ui.identity.CredentialManagerController
-import app.skerry.ui.connection.ConnectionController
-import app.skerry.ui.connection.ConnectionUiState
-import app.skerry.ui.connection.connectionSubtitle
-import app.skerry.ui.connection.toTarget
 import app.skerry.ui.host.HostManagerController
-import app.skerry.shared.ssh.TrustedCa
-import app.skerry.shared.ssh.TrustedCaStore
-import app.skerry.ui.known.KnownHostsController
-import app.skerry.ui.known.TrustedCaController
 import app.skerry.ui.session.SessionsController
 import app.skerry.ui.session.SessionView
 import app.skerry.ui.theme.SkerryTheme
 import app.skerry.ui.theme.ThemeMode
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import androidx.compose.runtime.snapshotFlow
-import kotlinx.coroutines.awaitCancellation
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withTimeoutOrNull
 import java.io.File
 import app.skerry.ui.vault.DesktopCorruptedScreen
 import app.skerry.ui.vault.DesktopCreateScreen
@@ -102,7 +26,6 @@ import app.skerry.ui.vault.DesktopResetScreen
 import app.skerry.ui.vault.DesktopUnlockScreen
 import app.skerry.ui.app.DesktopView
 import app.skerry.ui.mobile.MOBILE_PREVIEW_HOSTS
-import app.skerry.ui.design.DesignFonts
 import app.skerry.ui.design.LocalFonts
 import app.skerry.ui.metrics.MetricsAvailability
 import app.skerry.ui.metrics.PREVIEW_HOST_METRICS
@@ -114,19 +37,7 @@ import app.skerry.ui.mobile.MobileHostMonitorSheet
 import app.skerry.ui.app.MobileDesignState
 import app.skerry.ui.app.MobileRoute
 import app.skerry.ui.app.MobileTab
-import app.skerry.shared.host.VaultHostStore
-import app.skerry.shared.snippet.Snippet
-import app.skerry.shared.snippet.VaultSnippetStore
-import app.skerry.shared.vault.Credential
-import app.skerry.shared.vault.CredentialSecret
-import app.skerry.shared.vault.TrashStore
-import app.skerry.ui.theme.Skerry
-import app.skerry.ui.vault.TrashController
-import app.skerry.ui.vault.TrashList
 import app.skerry.ui.app.SettingsTab
-import app.skerry.ui.design.rememberMaterialSymbols
-import app.skerry.ui.design.rememberMono
-import app.skerry.ui.design.rememberUiFont
 
 /**
  * Offscreen render of the desktop design to PNG for visual review without a window/compositor.
@@ -329,621 +240,4 @@ private fun renderMobile(out: String, viewName: String, overlay: String, live: B
     scene.close()
     sessions?.disconnectAll() // detach fake session collectors before exit
     println("screenshot(mobile) → $out (${File(out).length()} bytes)")
-}
-
-/**
- * Points `user.home` at a throwaway directory holding a small demo tree, so the SFTP local pane
- * (real okio filesystem over `user.home`) loads instantly offscreen and shows a neutral path
- * instead of the developer's actual home. Rebuilt on each run; only used by the SFTP screenshot.
- */
-private fun seedFakeHome() {
-    val dir = File(System.getProperty("java.io.tmpdir"), "skerry-demo-home")
-    dir.deleteRecursively()
-    dir.mkdirs()
-    listOf("Projects", "Downloads", "Documents", ".ssh", ".config").forEach { File(dir, it).mkdirs() }
-    File(dir, "notes.md").writeText("# notes\n")
-    File(dir, "backup.tar.gz").writeText("demo\n")
-    File(dir, "deploy.log").writeText("ok\n")
-    File(dir, ".bashrc").writeText("export PATH\n")
-    System.setProperty("user.home", dir.absolutePath)
-}
-
-/** Design font provider for standalone screen renders, bypassing [DesktopDesignApp]. */
-/** Canned nginx.conf for the [FileEditorPanel] preview, over an in-memory source. */
-@Composable
-private fun EditorPreview() {
-    val scope = rememberCoroutineScope()
-    val controller = remember {
-        val text = """
-            server {
-                listen 443 ssl http2;
-                server_name skerry.app;
-
-                ssl_certificate     /etc/letsencrypt/live/skerry.app/fullchain.pem;
-                ssl_certificate_key /etc/letsencrypt/live/skerry.app/privkey.pem;
-
-                location / {
-                    proxy_pass http://127.0.0.1:8080;
-                    proxy_set_header Host ${'$'}host;
-                }
-            }
-        """.trimIndent()
-        val item = FileItem("nginx.conf", "/etc/nginx/sites-enabled/nginx.conf", FileItemType.File, text.length.toLong(), 0)
-        FileEditController(PreviewFileSource(item, text), item, readOnly = false, scope = scope).also { it.open() }
-    }
-    FileEditorScreen(controller, onClose = {}, modifier = Modifier.fillMaxSize())
-}
-
-/** In-memory single-file source for [EditorPreview]. */
-private class PreviewFileSource(private val item: FileItem, private val text: String) : FileContentBrowser {
-    override val label = "prod-web-01"
-    override suspend fun realpath(path: String) = path
-    override suspend fun list(path: String): List<FileItem> = emptyList()
-    override suspend fun mkdir(path: String) = Unit
-    override suspend fun delete(item: FileItem) = Unit
-    override suspend fun rename(from: String, to: String) = Unit
-    override suspend fun stat(path: String): FileItem = item
-    override suspend fun readFile(path: String, maxBytes: Long): ByteArray = text.encodeToByteArray()
-    override suspend fun writeFile(path: String, data: ByteArray) = Unit
-}
-
-/**
- * The assistant panel with a canned conversation, for visual review of the feed: an answer carrying
- * a command block, a block holding several separately runnable commands, and a destructive command
- * (which renders in the warning colour and asks twice before running).
- */
-@Composable
-private fun AssistantPreview() {
-    val controller = remember {
-        val scope = CoroutineScope(SupervisorJob() + Dispatchers.Unconfined)
-        var reply = ""
-        val provider = object : app.skerry.shared.ai.AiProvider {
-            override fun chat(request: app.skerry.shared.ai.AiChatRequest): Flow<app.skerry.shared.ai.AiDelta> =
-                flow { emit(app.skerry.shared.ai.AiDelta(reply)) }
-            override suspend fun close() {}
-        }
-        app.skerry.ui.ai.SessionAssistantController(
-            app.skerry.shared.ai.AiPolicy.Balanced,
-            settings = { app.skerry.shared.ai.AiSettings(apiKey = "sk-demo") },
-            providerFactory = { provider },
-            scope = scope,
-        ).apply {
-            reply = "The journal is the largest single consumer: **5.9 GiB** in `/var/log/journal`, " +
-                "on a filesystem that is 87% full. Docker layers take another 21 GiB.\n" +
-                "```bash\njournalctl --vacuum-size=500M\ndocker system prune -af\n```"
-            ask("Which service is eating the disk?", outputs = emptyList())
-            reply = "21 GiB in `/var/lib/docker`, of which 12.4 GiB are dangling layers built on this " +
-                "host. Pruning frees roughly 12 GiB without touching running containers.\n" +
-                "```\ndocker image prune -a --filter \"until=168h\"\n```"
-            ask("And the docker layers?", outputs = emptyList())
-            // One block holding several commands: what a model does when asked for a list, and each
-            // line has to stay separately runnable.
-            reply = "These are the ones worth having:\n```\ntop\nhtop\nvmstat 1 5\n```"
-            ask("Give me the basic load commands", outputs = emptyList())
-            // A genuinely destructive one, so the preview actually shows the warning colour and the
-            // two-click arming the doc above promises.
-            reply = "That frees the rotated logs, but it deletes them outright — there is no undo.\n" +
-                "```\nrm -rf /var/log/*.gz\n```"
-            ask("Can I just delete the rotated logs?", outputs = listOf("# ls /var/log", "# du -sh /var/log"))
-        }
-    }
-    androidx.compose.foundation.layout.Row(Modifier.fillMaxSize()) {
-        Box(Modifier.weight(1f).fillMaxSize())
-        app.skerry.ui.ai.AssistantPanel(
-            controller = controller,
-            terminal = null,
-            modelLabel = "local · Qwen2.5 Coder 7B",
-        )
-    }
-}
-
-/**
- * Standalone render of the Trash list ([TrashList]) over an in-memory vault with a few deletions,
- * for visual review of Settings -> Trash (and the identical Android screen). The clock is fixed so
- * the "days left" column is stable between runs.
- */
-@Composable
-private fun TrashPreview() {
-    val controller = remember {
-        val vault = InMemoryVault()
-        var clock = 1_800_000_000_000L
-        val trash = TrashStore(vault, now = { clock })
-        VaultHostStore(vault, trash = trash).apply {
-            put(Host(id = "h-1", label = "staging-web", address = "10.0.0.14", port = 22, username = "deploy"))
-            remove("h-1")
-        }
-        clock -= 9L * 24 * 60 * 60 * 1000
-        CredentialStore(vault, trash).apply {
-            put(Credential("c-1", "db-admin", CredentialSecret.Password("hunter2")))
-            remove("c-1")
-        }
-        clock -= 17L * 24 * 60 * 60 * 1000
-        VaultSnippetStore(vault, trash).apply {
-            put(Snippet("s-1", "tail syslog", "tail -f /var/log/syslog"))
-            remove("s-1")
-        }
-        TrashController(trash, now = { 1_800_000_000_000L })
-    }
-    Box(Modifier.fillMaxSize().background(Skerry.colors.surfaceDeep).padding(32.dp)) {
-        TrashList(controller)
-    }
-}
-
-@Composable
-private fun GateScreenPreview(body: @Composable () -> Unit) {
-    val fonts = DesignFonts(
-        ui = rememberUiFont(),
-        mono = rememberMono(),
-        symbols = rememberMaterialSymbols(),
-    )
-    CompositionLocalProvider(LocalFonts provides fonts) { body() }
-}
-
-/**
- * In-memory host catalog with demo profiles, for the offscreen render of a live sidebar only.
- * [boundCredentialId] (if given) is attached to a pair of hosts so Vault shows "Used by hosts".
- */
-private fun seededHosts(boundCredentialId: String? = null): HostManagerController {
-    val store = object : HostStore {
-        private val items = LinkedHashMap<String, Host>()
-        override fun all(): List<Host> = items.values.toList()
-        override fun put(host: Host) { items[host.id] = host }
-        override fun remove(id: String) { items.remove(id) }
-        override fun reorder(transform: (List<Host>) -> List<Host>) {
-            val updated = transform(items.values.toList())
-            items.clear()
-            updated.forEach { items[it.id] = it }
-        }
-    }
-    listOf(
-        Host("h1", "prod-web-01", "192.168.1.45", 22, "root", "Production", credentialId = boundCredentialId, tags = listOf("prod", "web")),
-        Host("h2", "db-master", "192.168.1.50", 22, "root", "Production", credentialId = boundCredentialId, tags = listOf("prod", "db")),
-        Host("h3", "homelab-pi", "10.0.0.12", 22, "pi", "Homelab", tags = listOf("docker")),
-        Host("h4", "vps-edge", "vps.example.com", 2222, "deploy", null, tags = listOf("edge")),
-        // Remote desktops: their own section in the shell, so the demo catalog seeds both kinds.
-        Host("h5", "lab-desktop", "10.0.0.30", 5900, "", "Homelab", connectionType = ConnectionType.VNC),
-        Host("h6", "win-bench", "10.0.0.31", 5901, "", null, connectionType = ConnectionType.VNC, tags = listOf("lab")),
-    ).forEach(store::put)
-    var seq = 0
-    return HostManagerController(store) { "gen-${seq++}" }
-}
-
-/**
- * Seeds the vault with keychain secrets ([CredentialManagerController]) over an in-memory vault so
- * the offscreen render shows a live Vault section (key/password/certificate cards, used-by-hosts)
- * with real components and no files/master password. One ed25519 key is generated by a real
- * [SshKeyGenerator]; the first secret is attached to the demo hosts by credentialId.
- */
-private fun seededVault(keyGenerator: SshKeyGenerator): CredentialManagerController {
-    val vault = InMemoryVault()
-    var credSeq = 0
-    val credentials = CredentialManagerController(CredentialStore(vault)) { "cred-${credSeq++}" }
-
-    val key = keyGenerator.generate(SshKeyType.ED25519, comment = "alice@skerry")
-    credentials.save(CredentialDraft(label = "work-laptop", kind = CredentialKind.PRIVATE_KEY, privateKeyPem = key.privateKeyPem))
-    credentials.save(CredentialDraft(label = "db-admin", kind = CredentialKind.PASSWORD, password = "hunter2"))
-    credentials.save(
-        CredentialDraft(label = "prod-access", kind = CredentialKind.CERTIFICATE, privateKeyPem = SEED_CERT_KEY, certificate = SEED_CERT),
-    )
-
-    return credentials
-}
-
-// NOT_A_SECRET: a throwaway ed25519 key generated for offscreen test seeding; not used in the
-// production build (desktopMain Screenshot only), grants access to nothing. Marker so
-// gitleaks/trufflehog don't false-positive on it.
-// Throwaway ed25519 certificate (CA-signed, principals alice/deploy), only for seeding the offscreen
-// render of Vault -> Certificates with real components, no files/master password. Same values as
-// CertificateFixtures (shared/desktopTest), which is the source of truth; update both if regenerated.
-private val SEED_CERT_KEY = """
-    -----BEGIN OPENSSH PRIVATE KEY-----
-    b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAAAMwAAAAtzc2gtZW
-    QyNTUxOQAAACCHmK+eOLE/3SmTEHz2mQerUTWuK10g2yXsCeRmqBhDJwAAAJCTquJek6ri
-    XgAAAAtzc2gtZWQyNTUxOQAAACCHmK+eOLE/3SmTEHz2mQerUTWuK10g2yXsCeRmqBhDJw
-    AAAECj4nk0xG00zyQDEYjZzkq4DYaRGzTDQCa722CqWQsnKIeYr544sT/dKZMQfPaZB6tR
-    Na4rXSDbJewJ5GaoGEMnAAAADGFsaWNlQHNrZXJyeQE=
-    -----END OPENSSH PRIVATE KEY-----
-""".trimIndent() + "\n"
-
-private const val SEED_CERT =
-    "ssh-ed25519-cert-v01@openssh.com AAAAIHNzaC1lZDI1NTE5LWNlcnQtdjAxQG9wZW5zc2guY29tAAAAIJ/XTmChh23PUo43PsVebZVnBUh9yVb7r8UgCo6MD2XGAAAAIIeYr544sT/dKZMQfPaZB6tRNa4rXSDbJewJ5GaoGEMnAAAAAAAAACoAAAABAAAAE3NrZXJyeS10ZXN0QGVkMjU1MTkAAAATAAAABWFsaWNlAAAABmRlcGxveQAAAABlkgCAAAAAAHhh+AAAAAAAAAAAggAAABVwZXJtaXQtWDExLWZvcndhcmRpbmcAAAAAAAAAF3Blcm1pdC1hZ2VudC1mb3J3YXJkaW5nAAAAAAAAABZwZXJtaXQtcG9ydC1mb3J3YXJkaW5nAAAAAAAAAApwZXJtaXQtcHR5AAAAAAAAAA5wZXJtaXQtdXNlci1yYwAAAAAAAAAAAAAAMwAAAAtzc2gtZWQyNTUxOQAAACDGkIM6oT/mc8hunaUIY1avJGKsnfJB6yboLBsENiQ0kAAAAFMAAAALc3NoLWVkMjU1MTkAAABAwycZAnZtpvGb6wZDhWCcA6sa4Lz7sieexLCRkC7VNcZj23iiqej1B135atUIc0G7yR/g/TIzACfk2G3DHOYLAA== alice@skerry"
-
-/** Trivial unencrypted in-memory vault, for offscreen identity seeding only (not for the app). */
-private class InMemoryVault : Vault {
-    private val records = LinkedHashMap<String, VaultRecord>()
-    private val payloads = LinkedHashMap<String, ByteArray>()
-
-    override fun exists(): Boolean = true
-    override val isUnlocked: Boolean = true
-    override fun create(password: CharArray) = Unit
-    override fun unlock(password: CharArray): UnlockResult = UnlockResult.Success
-    override fun unlockWithDataKey(dataKey: DataKey): UnlockResult = UnlockResult.Success
-    override fun exportDataKey(): DataKey? = null
-    override fun adoptDataKey(newDataKey: DataKey, password: CharArray): Boolean = false
-    override fun lock() = Unit
-    override fun reset() { records.clear(); payloads.clear() }
-    override fun records(): List<VaultRecord> = records.values.filterNot { it.deleted }
-    override fun syncMeta(): SyncMeta? = null
-    override fun mergeRemote(remote: List<VaultRecord>): MergeResult = MergeResult.EMPTY
-    override fun openPayload(id: String): ByteArray? = payloads[id]
-    override fun put(id: String, type: RecordType, payload: ByteArray) {
-        payloads[id] = payload
-        records[id] = VaultRecord(id, type, version = 1, updatedAt = "", deviceId = "screenshot", deleted = false, blob = ByteArray(0))
-    }
-    override fun remove(id: String) {
-        payloads.remove(id)
-        records[id]?.let { records[id] = it.copy(deleted = true) }
-    }
-    override fun changePassword(oldPassword: CharArray, newPassword: CharArray): Boolean = true
-    override fun verifyPassword(password: CharArray): Boolean = true
-}
-
-/** A short canned recording for the player render (`view=Player`). */
-private fun seededCast(): Asciicast = Asciicast(
-    columns = 138,
-    rows = 30,
-    title = "deploy.cast",
-    events = listOf(
-        CastEvent(0.0, "\u001b[36mroot@prod-web-01\u001b[0m:~# ./deploy.sh\r\n"),
-        CastEvent(0.2, "  \u001b[32m✓\u001b[0m build      12.4s\r\n"),
-        CastEvent(0.4, "  \u001b[32m✓\u001b[0m tests      31.2s\r\n"),
-        CastEvent(0.6, "  \u001b[32m✓\u001b[0m rollout     4.8s\r\n"),
-        CastEvent(0.8, "\r\ndeployed \u001b[1mv0.1.9\u001b[0m to 3 nodes\r\n"),
-        // Full-width rule: shows at a glance whether the recording fills the pane edge to edge.
-        CastEvent(1.0, "\u001b[2m" + "\u2500".repeat(138) + "\u001b[0m"),
-    ),
-)
-
-/**
- * Tunnel manager over an in-memory store and the fake transport, so the offscreen Ports render
- * shows the live table (one active forward with its browser link) and a real service scan against
- * [FakeSshConnection]'s canned `ss` output.
- */
-private fun seededTunnels(hosts: HostManagerController): TunnelManager {
-    val store = object : TunnelStore {
-        private val entries = mutableListOf<Tunnel>()
-        override fun all(): List<Tunnel> = entries.toList()
-        override fun put(tunnel: Tunnel) {
-            val i = entries.indexOfFirst { it.id == tunnel.id }
-            if (i >= 0) entries[i] = tunnel else entries += tunnel
-        }
-        override fun remove(id: String) { entries.removeAll { it.id == id } }
-    }
-    val hostIds = hosts.hosts.map { it.id }
-    fun hostAt(i: Int) = hostIds.getOrElse(i) { hostIds.first() }
-    store.put(Tunnel("t1", "web tunnel", hostAt(0), TunnelDirection.Local, "127.0.0.1", 8080, "10.0.0.5", 80, autostart = true))
-    store.put(Tunnel("t2", "app callback", hostAt(1), TunnelDirection.Remote, "0.0.0.0", 9000, "localhost", 3000))
-    store.put(Tunnel("t3", "socks", hostAt(2), TunnelDirection.Dynamic, "127.0.0.1", 1080, null, null, autostart = true))
-    var next = 0
-    val manager = TunnelManager(
-        store = store,
-        transport = fakeTransport(),
-        resolve = { hostId ->
-            val host = hosts.find(hostId)
-            if (host == null) {
-                TunnelResolution.Unavailable(TunnelUnavailable.HostNotFound)
-            } else {
-                TunnelResolution.Ready(host.toTarget(), SshAuth.Password("demo"))
-            }
-        },
-        scope = CoroutineScope(SupervisorJob() + Dispatchers.Default),
-    ) { "seed-${next++}" }
-    manager.activate("t1")
-    // The scene renders one frame at t=0, so everything the one-second telemetry poll feeds — the
-    // traffic column, the throughput sparkline — would be empty. Wait for the forward, then drive
-    // the poll by hand to fill the window.
-    runBlocking { withTimeoutOrNull(2000) { while (manager.find("t1")?.status !is TunnelStatus.Active) delay(10) } }
-    repeat(TunnelTelemetry.HISTORY_CAPACITY) { manager.pollTelemetry() }
-    // -Dskerry.screenshot.portsScan=true renders the Services panel instead of the tunnel editor
-    // (the offscreen scene can't press Find services / Scan itself).
-    if (System.getProperty("skerry.screenshot.portsScan", "false").toBoolean()) manager.services.scan(hostAt(0))
-    return manager
-}
-
-/**
- * Known-hosts manager with demo keys and one unresolved key-change event, so the offscreen render
- * shows a live table (firstSeen/Verified), a warning, and a fingerprint comparison panel with real
- * components ([KnownHostsController] -> [KnownHostsView]). In-memory, no files.
- */
-/**
- * Trusted certificate authorities for the offscreen render of the known-hosts screen: two entries
- * over an in-memory store, so the CA section shows real rows rather than its empty state.
- */
-private fun seededTrustedCas(): TrustedCaController {
-    val store = object : TrustedCaStore {
-        private val items = mutableListOf(
-            TrustedCa("ca-1", "*.prod.example.com", "ssh-ed25519", "AAAA", "SHA256:Qz8kR2mVpL4tXw9nB7sJ1dF0", "Production CA", "2026-05-02T10:00:00Z"),
-            TrustedCa("ca-2", "*.staging.example.com,!admin.staging.example.com", "ecdsa-sha2-nistp256", "AAAA", "SHA256:7hT4bN0xZq2wE5rY8uI3oP6a", "Staging CA", "2026-06-18T10:00:00Z"),
-        )
-        override fun all() = items.toList()
-        override fun put(ca: TrustedCa) {
-            items.removeAll { it.id == ca.id }
-            items += ca
-        }
-        override fun remove(id: String) {
-            items.removeAll { it.id == id }
-        }
-    }
-    return TrustedCaController(store, { null }, newId = { "ca-x" }, now = { "2026-06-22T12:00:00Z" })
-}
-
-private fun seededKnownHosts(): KnownHostsController {
-    val ed = "ssh-ed25519"
-    val store = object : KnownHostsStore {
-        private val items = mutableListOf(
-            KnownHost("prod-web-01", 22, ed, "SHA256:8c3F1a2bQzABCDEFGHIJKLMNpK9R", "2026-01-12T09:00:00Z"),
-            KnownHost("db-master", 22, ed, "SHA256:2dE7bLm4xRABCDEFGHIJKLMNwQ1z", "2026-01-12T09:05:00Z"),
-            KnownHost("nas-truenas", 22, ed, "SHA256:9aB0cTn2wE4rXp1kLm7sQ8vZabcd", "2026-03-04T18:30:00Z"),
-            KnownHost("homelab-pi", 22, "ssh-rsa", "SHA256:5fG1hKp8sXYZ0123456789vB3nqrst", "2026-02-02T11:15:00Z"),
-        )
-        override fun all() = items.toList()
-        override fun add(host: KnownHost) { items += host }
-        override fun replace(host: KnownHost) {
-            items.removeAll { it.host == host.host && it.port == host.port && it.keyType == host.keyType }
-            items += host
-        }
-        override fun remove(host: String, port: Int, keyType: String) {
-            items.removeAll { it.host == host && it.port == port && it.keyType == keyType }
-        }
-    }
-    val mismatches = object : HostKeyMismatchStore {
-        private val items = mutableListOf(
-            HostKeyMismatch("nas-truenas", 22, ed, "SHA256:9aB0cTn2wE4rXp1kLm7sQ8vZabcd", "SHA256:Kp3xQ9zR1tWv7nB4mL0sJ2dFefgh", "2026-06-22T08:00:00Z"),
-        )
-        override fun all() = items.toList()
-        override fun record(mismatch: HostKeyMismatch) {
-            items.removeAll { it.host == mismatch.host && it.port == mismatch.port && it.keyType == mismatch.keyType }
-            items += mismatch
-        }
-        override fun clear(host: String, port: Int, keyType: String) {
-            items.removeAll { it.host == host && it.port == port && it.keyType == keyType }
-        }
-    }
-    return KnownHostsController(store, mismatches) { "2026-06-22T12:00:00Z" }
-}
-
-/**
- * Live AI controller for offscreen renders of AI settings (desktop tab and mobile screen): a fake
- * provider with a canned reply, the first catalog model marked "installed", a BYOK key filled in,
- * quick-chat seeded with one exchange. Provider is set via `-Dskerry.screenshot.aiProvider`
- * (CLOUD/DEVICE/OFF, default CLOUD); OFF renders the "AI disabled" state.
- */
-/**
- * Live update-notice controller for offscreen renders: the "check for updates" toggle is on, and
- * `-Dskerry.screenshot.updateAvailable=true` makes the canned newer release visible (status-bar
- * item, About notice, More-row subtitle); without the flag the check finds nothing.
- */
-private fun seededUpdates(): app.skerry.ui.update.UpdateNoticeController {
-    val scope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
-    val available = System.getProperty("skerry.screenshot.updateAvailable", "false").toBoolean()
-    var settings = app.skerry.shared.update.UpdateSettings()
-    val controller = app.skerry.ui.update.UpdateNoticeController(
-        initialSettings = settings,
-        persist = { settings = it },
-        check = {
-            if (available) {
-                app.skerry.shared.update.AvailableUpdate("0.2.0", "https://github.com/SeCherkasov/SkerrySSH/releases/tag/v0.2.0")
-            } else {
-                null
-            }
-        },
-        scope = scope,
-        reload = { settings },
-    )
-    controller.refresh()
-    return controller
-}
-
-private fun seededAi(): app.skerry.ui.ai.AiAssistantController {
-    val scope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
-    val kind = runCatching {
-        app.skerry.shared.ai.AiProviderKind.valueOf(System.getProperty("skerry.screenshot.aiProvider", "CLOUD"))
-    }.getOrDefault(app.skerry.shared.ai.AiProviderKind.CLOUD)
-    val first = app.skerry.shared.ai.local.LocalModelCatalog.models.first()
-    var settings = app.skerry.shared.ai.AiSettings(apiKey = "sk-demo", provider = kind, localModelId = first.id)
-    val fakeProvider = object : app.skerry.shared.ai.AiProvider {
-        override fun chat(request: app.skerry.shared.ai.AiChatRequest): Flow<app.skerry.shared.ai.AiDelta> = flow {
-            emit(app.skerry.shared.ai.AiDelta("Use scp: scp file.txt user@host:/path/ — it copies over SSH with the same credentials."))
-        }
-        override suspend fun close() {}
-    }
-    val controller = app.skerry.ui.ai.AiAssistantController(
-        initialSettings = settings,
-        persist = { settings = it },
-        providerFactory = { fakeProvider },
-        scope = scope,
-        reload = { settings },
-        localInstalled = { it.id == first.id },
-        models = app.skerry.ui.ai.LocalModelController(
-            installed = { it.id == first.id },
-            fetch = { flow {} },
-            remove = {},
-            scope = scope,
-        ),
-    )
-    if (controller.enabled) controller.ask("how do I copy a file to the server?")
-    return controller
-}
-
-/**
- * Session manager over a fake transport ([fakeTransport]) with one open tab to the first host, so
- * the offscreen render shows a live terminal/toolbar/tabs with real components
- * ([SessionsController] -> [ConnectionController] -> [TerminalScreen]), no network.
- */
-private fun seededSessions(hosts: HostManagerController): SessionsController {
-    val scope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
-    var n = 0
-    val sessions = SessionsController(newId = { "s${n++}" }, controllerFactory = { ConnectionController(fakeTransport(), scope) })
-    // Empty password: the fake transport ignores auth (see FakeConnection); there's no real handshake.
-    val h = hosts.hosts.first()
-    sessions.open(h.id, h.label, h.connectionSubtitle(), h.toTarget(), SshAuth.Password(""))
-    hosts.hosts.getOrNull(1)?.let { h2 -> sessions.open(h2.id, h2.label, h2.connectionSubtitle(), h2.toTarget(), SshAuth.Password("")) }
-    sessions.activate(sessions.tabs.first().id)
-    // Seeds port forwards on the active session for a live Tunnels tab screenshot: waits for the
-    // fake connection to come up (connect is async), then raises -L/-R/-D the same way the UI does
-    // (PortForwardController). The fake forward is Active immediately.
-    // Split grid for the pane screenshots: -Dskerry.screenshot.panes=4 fills the active tab with
-    // that many connected panes and turns synchronized input on, which is what the work bar's split
-    // title is there to show.
-    val panes = System.getProperty("skerry.screenshot.panes", "1").toIntOrNull() ?: 1
-    val tabId = sessions.tabs.first().id
-    repeat((panes - 1).coerceAtLeast(0)) { i ->
-        // Loud rather than a quietly under-filled picture: a grid that refuses another pane
-        // means the requested -Dskerry.screenshot.panes cannot be rendered at all.
-        val paneId = checkNotNull(sessions.addPane(tabId)) { "pane $i of $panes refused: grid is full" }
-        val host = hosts.hosts[(i + 1) % hosts.hosts.size]
-        sessions.connectPane(tabId, paneId, host.id, host.label, host.connectionSubtitle(), host.toTarget(), SshAuth.Password(""))
-    }
-    if (panes > 1) {
-        sessions.toggleSyncInput(tabId)
-        sessions.focusPane(tabId, sessions.tabs.first().panes.first().id)
-    }
-    val ctrl = sessions.tabs.first().focusedPane.controller
-    scope.launch {
-        // uiState is Compose snapshot state; waits for the transition to Connected via snapshotFlow
-        // (not a busy-spin) to properly subscribe to the snapshot system without spinning the CPU.
-        snapshotFlow { ctrl.uiState }.first { it is ConnectionUiState.Connected }
-        val pf = ctrl.openPortForwards()
-        pf.addLocal(bindPort = 0, destHost = "10.0.0.5", destPort = 80)
-        pf.addRemote(bindPort = 9000, destHost = "localhost", destPort = 3000)
-        pf.addDynamic(bindPort = 1080)
-    }
-    return sessions
-}
-
-private val SEEDED_SS_OUTPUT = """
-    State  Recv-Q Send-Q Local Address:Port Peer Address:Port Process
-    LISTEN 0      128            0.0.0.0:22        0.0.0.0:*    users:(("sshd",pid=640,fd=3))
-    LISTEN 0      511            0.0.0.0:80        0.0.0.0:*    users:(("nginx",pid=901,fd=6))
-    LISTEN 0      4096         127.0.0.1:5432      0.0.0.0:*    users:(("postgres",pid=812,fd=5))
-    LISTEN 0      511          127.0.0.1:6379      0.0.0.0:*    users:(("redis-server",pid=733,fd=7))
-    LISTEN 0      128             0.0.0.0:8080     0.0.0.0:*    users:(("java",pid=1204,fd=41))
-""".trimIndent()
-
-/** Fake SSH transport: the shell emits a canned banner+listing, then hangs until cancelled. */
-private fun fakeTransport(): SshTransport = object : SshTransport {
-    override suspend fun connect(target: SshTarget, auth: SshAuth): SshConnection = FakeConnection(target)
-}
-
-private class FakeConnection(private val target: SshTarget) : SshConnection {
-    override val isConnected: Boolean = true
-    override val cipher: String? = "chacha20-poly1305@openssh.com"
-
-    // Realistic METRICS_COMMAND output so the offscreen render of the info panel shows a live
-    // monitor (CPU/memory/swap, network, filesystems, top processes, host facts) instead of "..."
-    // placeholders. The network counters advance per poll so the rates and their sparklines have
-    // something to draw.
-    private var polls = 0
-
-    override suspend fun exec(command: String): ExecResult {
-        // Service discovery asks a different question than the metrics poll; answer it with canned
-        // `ss -ltnp` output so the offscreen Ports render shows a real scan result.
-        if (command == SERVICE_SCAN_COMMAND) return ExecResult(0, SEEDED_SS_OUTPUT, "")
-        val tick = polls++
-        return ExecResult(
-            exitCode = 0,
-            stdout = """
-                cpu  100 0 100 800 0 0 0 0
-                cpu  168 0 132 900 0 0 0 0
-                @MEM
-                Mem:     4000000000  2100000000  1000000000
-                Swap:    2000000000   210000000  1790000000
-                @DISK
-                Filesystem     1024-blocks      Used Available Capacity Mounted on
-                /dev/sda1         51475068  42000000   6900000      87% /
-                tmpfs               400000      1000    399000       1% /run
-                /dev/sda2        209715200 120000000  78000000      62% /var
-                @NET
-                    lo: 1000000 100 0 0 0 0 0 0 1000000 100
-                  eth0: ${'$'}{4_000_000 + tick * 380_000} 500 0 0 0 0 0 0 ${'$'}{1_200_000 + tick * 95_000} 200
-                @PROC
-                2481  12.4  8.1 postgres
-                 991   4.2  2.3 nginx
-                1204   1.8  5.7 node
-                @UPTIME
-                372765.42 1488907.15
-                @LOAD
-                0.42 0.51 0.48 1/512 28931
-                @OS
-                PRETTY_NAME="Ubuntu 22.04.4 LTS"
-                @KERNEL
-                Linux 5.15.0-105-generic x86_64
-                @CPU
-                4
-            """.trimIndent(),
-            stderr = "",
-        )
-    }
-
-    override suspend fun openShell(size: PtySize, term: String): ShellChannel = FakeChannel(target)
-    override suspend fun openSftp(): SftpClient = FakeSftpClient()
-    override suspend fun forwardLocal(spec: LocalForwardSpec): PortForward = FakePortForward(if (spec.bindPort != 0) spec.bindPort else 50080)
-    override suspend fun forwardRemote(spec: RemoteForwardSpec): PortForward = FakePortForward(if (spec.bindPort != 0) spec.bindPort else 9000)
-    override suspend fun forwardDynamic(spec: DynamicForwardSpec): PortForward = FakePortForward(if (spec.bindPort != 0) spec.bindPort else 1080)
-    override suspend fun disconnect() {}
-}
-
-/** Fake forward: immediately "active", port echoed from the spec, for the offscreen tunnels table. */
-private class FakePortForward(override val boundPort: Int) : PortForward {
-    override val isActive: Boolean = true
-    override var isPaused: Boolean = false
-        private set
-
-    // Counters grow with every read, on a small wave, so the traffic column and the throughput
-    // sparkline render with something in them. A constant would draw a flat line and read as a bug.
-    private var reads = 0L
-    override val bytesUp: Long get() = 46_000L * ++reads + (reads % 11) * 2_400
-    override val bytesDown: Long get() = 138_000L * reads + (reads % 13) * 5_800
-    override suspend fun pause() { isPaused = true }
-    override suspend fun resume() { isPaused = false }
-    override suspend fun close() = Unit
-}
-
-private class FakeChannel(target: SshTarget) : ShellChannel {
-    private val prompt = "${target.username}@${target.host.substringBefore('.')}:~# "
-    private val banner =
-        "Last login: Sat Jun 21 14:22:10 2026 from 10.0.0.15\r\n" +
-            "$prompt" + "ls -la\r\n" +
-            "total 24\r\n" +
-            "drwxr-xr-x  5 root root 4096 Jun 21 14:02 app\r\n" +
-            "drwxr-xr-x  2 root root 4096 Jun 21 09:11 deploy\r\n" +
-            "-rw-r--r--  1 root root  812 Jun 20 23:40 backup.tar.gz\r\n" +
-            "$prompt" + "df -h /\r\n" +
-            "Filesystem      Size  Used Avail Use% Mounted on\r\n" +
-            "/dev/sda1        50G   42G  5.2G  87% /\r\n" +
-            prompt
-
-    override val isOpen: Boolean = true
-    override val output: Flow<ByteArray> = flow {
-        emit(banner.encodeToByteArray())
-        awaitCancellation()
-    }
-
-    override suspend fun write(data: ByteArray) {}
-    override suspend fun resize(size: PtySize) {}
-    override suspend fun close() {}
-}
-
-/** Fake SFTP client with a canned `/var/www` listing, for the offscreen render of a live panel. */
-private class FakeSftpClient : SftpClient {
-    private val listing = listOf(
-        SftpEntry("html", "/var/www/html", SftpEntryType.Directory, 4096, 0, 0b111_101_101),
-        SftpEntry("releases", "/var/www/releases", SftpEntryType.Directory, 4096, 0, 0b111_101_101),
-        SftpEntry("nginx.conf", "/var/www/nginx.conf", SftpEntryType.File, 3174, 0, 0b110_100_100),
-        SftpEntry("robots.txt", "/var/www/robots.txt", SftpEntryType.File, 112, 0, 0b110_100_100),
-        SftpEntry("deploy.sh", "/var/www/deploy.sh", SftpEntryType.File, 1843, 0, 0b111_101_101),
-    )
-
-    override suspend fun list(path: String): List<SftpEntry> = listing
-    override suspend fun stat(path: String): SftpEntry? = null
-    override suspend fun realpath(path: String): String = "/var/www"
-    override suspend fun read(path: String, maxBytes: Long): ByteArray = ByteArray(0)
-    override suspend fun write(path: String, data: ByteArray) = Unit
-    override suspend fun download(remotePath: String, localPath: String, onProgress: SftpProgress) = Unit
-    override suspend fun upload(localPath: String, remotePath: String, onProgress: SftpProgress) = Unit
-    override suspend fun mkdir(path: String) = Unit
-    override suspend fun remove(path: String) = Unit
-    override suspend fun rmdir(path: String) = Unit
-    override suspend fun rename(from: String, to: String) = Unit
-    override suspend fun close() = Unit
 }
